Initial commit. Working blink test with Uno r3. Working on flashing methods for Claude Code.
Claude Code bug: Claude Code gets hung when working with flashing process as it is waiting for a state change that it doesn't understand.

flash-external launches flasher externally from Claude Code in terminal and then closes terminal when the process is finished. Claude detects this but after double-time of completed flash. Working on monitoring/state detection to overcome this and create seamless dev experience.

Will expand flash-external to include CMD/PS in the future for Win devices.

Continuing to solve Claude Code bug for better experience.